requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.22
arxiv>=2.1.0
tiktoken>=0.5.2
python-dotenv>=1.0.0
